Conditionally unpack an error_id from a thrown system_error
Previously, if a std::system_error was caught and passed to
unpack_error_id, then the 'exc.code()' would be unconditionally
converted to an error_id, even if the code did not correspond
to an actual error_id value. This effectively prevents reliable
use of std::system_error as an exception type.

This tweaks to only consider the system_error to be an error_id
if the code() of that error matches is_error_id().