Check for handle_more when finding a handler
If a handler returns a more_handlers_result, then at least one of
those inner handlers must also be invocable with the current error. If
none of those handlers are callable, then the wrapping handler is not
viable either. This constraint also applies to "any error" handlers that
return a handle_more.